Refactor, re-organize and rename
