+++ pkg/db/src/implementations/version-1.2/books.ts
@@ -146,7 +146,6 @@ const createSearchIndex = (books: BookEntry[]) => {
 	index.addIndex("authors");
 	index.addIndex("publisher");
 	index.addIndex("editedBy");
-	index.addIndex("committedAt");
 
 	index.addDocuments(books);
